requirements.txt
Flask==1.0.2
dateutils==0.6.6
python-dateutil==2.7.5
python-dotenv==0.10.1
requests==2.21.0
typing==3.6.6
unittest2==1.1.0
cloudant==2.10.2
ibm-cos-sdk==2.4.2
